Clean-up Genetechs website
Starting with the redesign I encountered some unnecessary code and
not-so-semantic parts in the HTML. That's why I decided to do a clean-up
first, so that our code will be clean and fast.

I made some quick changes on the design as well, but those were only to
make it a bit more consistent with the rest of our products so far (our
printed work and web-application). The rest of the re-design can start
now.